Moved some functions from Game to Framwork
